+++ hata/discord/message/message.py
@@ -573,7 +573,7 @@ class Message(DiscordEntity, immortal=True):
         
         interaction_data = data.get('interaction', None)
         if (interaction_data is not None):
-            interaction = MessageInteraction(interaction_data)
+            interaction = MessageInteraction(interaction_data, guild_id)
             try_resolve_interaction_message(self, interaction)
             
             _set_message_field(
@@ -662,7 +662,7 @@ class Message(DiscordEntity, immortal=True):
         if update_interaction:
             interaction_data = data.get('interaction', None)
             if (interaction_data is not None):
-                interaction = MessageInteraction(interaction_data)
+                interaction = MessageInteraction(interaction_data, self.guild_id)
                 try_resolve_interaction_message(self, interaction)
                 
                 _set_message_field(